# src/server.ts
import express from "express";
import { prisma } from "../lib/prisma.js";
import swaggerUi from "swagger-ui-express";

import swaggerDocument from "../swagger.json" with { type: "json" };

const port = 3000;
const app = express();
app.use(express.json());

app.get("/movies", async (_, res) => {
   const movies = await prisma.movie.findMany({
      orderBy: {
         title: "asc",
      },
      include: {
         genres: true,
         languages: true,
      },
   });
   res.json(movies);
});

app.post("/movies", async (req, res) => {
   const { id, title, genre_id, language_id, oscar_count, release_date } =
      req.body;

   try {
      const movieWithSameTitle = await prisma.movie.findFirst({
         where: {
            title: { equals: title, mode: "insensitive" },
         },
      });

      if (movieWithSameTitle) {
         return res
            .status(409)
            .send({ message: "Já existe um filme com esse título" });
      }

      await prisma.movie.create({
         data: {
            id,
            title,
            genre_id,
            language_id,
            oscar_count,
            release_date: new Date(release_date),
         },
      });
   } catch (error) {
      res.status(500).send({ message: "Falha ao cadastrar um filme" });
   }

   res.status(201).send();
});

app.put("/movies/:id", async (req, res) => {
   const id = Number(req.params.id);

   try {
      const movie = await prisma.movie.findUnique({
         where: {
            id,
         },
      });

      if (!movie) {
         return res.status(404).send({ message: "Filme não encontrado" });
      }

      const data = { ...req.body };

      data.release_date = data.release_date
         ? new Date(data.release_date)
         : undefined;

      await prisma.movie.update({
         where: {
            id,
         },
         data,
      });
   } catch (error) {
      return res
         .status(500)
         .send({ message: "Falha ao atualizar o registro do filme" });
   }

   res.status(200).send();
});

app.delete("/movies/:id", async (req, res) => {
   const id = Number(req.params.id);

   try {
      const movie = await prisma.movie.findUnique({
         where: {
            id,
         },
      });

      if (!movie) {
         return res.status(404).send({ message: "Filme não encontrado" });
      }

      await prisma.movie.delete({
         where: {
            id,
         },
      });
   } catch (error) {
      return res.status(500).send({ message: "Falha ao remover um filme" });
   }

   res.status(200).send();
});

app.get("/movies/:genreName", async (req, res) => {
   const genreName = req.params.genreName;

   try {
      const moviesFilteredByGenreName = await prisma.movie.findMany({
         include: {
            genres: true,
            languages: true,
         },
         where: {
            genres: {
               name: {
                  equals: genreName,
                  mode: "insensitive",
               },
            },
         },
      });

      if (!moviesFilteredByGenreName.length) {
         return res
            .status(404)
            .send({ message: "Não foi encontrado um filme com esse gênero" });
      }

      res.status(200).send(moviesFilteredByGenreName);
   } catch (error) {
      return res
         .status(500)
         .send({ message: "Falha ao filtar o filme pelo gênero" });
   }
});

app.use("/docs", swaggerUi.serve, swaggerUi.setup(swaggerDocument));

app.listen(port, () => {
   console.log(`Servidor em execução na porta ${port}`);
});